+++ guang-cli/packages/cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @guang-cli/cli
 
+## 0.1.0
+
+### Minor Changes
+
+- generate command
+
+### Patch Changes
+
+- Updated dependencies
+  - @guang-cli/generate@1.2.0
+
 ## 0.0.7
 
 ### Patch Changes