+++ guang-cli/packages/template-vue/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @guang-cli/template-vue
 
+## 1.2.2
+
+### Patch Changes
+
+- generate command
+
+## 1.2.1
+
+### Patch Changes
+
+- generate + config
+
 ## 1.2.0
 
 ### Minor Changes